Implemented and tweaked API controllers.

+++ FribergFastigheter.Server/Controllers/BrokerFirmApi/BrokerHousingController.cs
@@ -53,36 +53,49 @@ namespace FribergFastigheter.Server.Controllers.BrokerFirmApi
             _configuration = configuration;
         }
 
-        #endregion
+		#endregion
 
-        #region ApiEndPoints
+		#region ApiEndPoints
 
-        /// <summary>
-        /// An API endpoint for deleting housing objects. 
-        /// </summary>
-        /// <param name="id">The ID of the housing object to delete.</param>
-        /// <!-- Author: Jimmie -->
-        /// <!-- Co Authors: -->
-        [HttpDelete("{id:int}")]
-        public async Task<ActionResult> Delete(int id)
+		/// <summary>
+		/// An API endpoint for deleting housing objects. 
+		/// </summary>
+		/// <param name="id">The ID of the housing object to delete.</param>
+		/// <param name="brokerFirmId">The ID of the broker firm associated with the housing.</param>
+		/// <!-- Author: Jimmie -->
+		/// <!-- Co Authors: -->
+		[HttpDelete("{id:int}")]
+        public async Task<ActionResult> Delete(int id, [Required] int brokerFirmId)
         {
+            var housing = await _housingRepository.GetHousingByIdAsync(id);
+
+            if (!await _housingRepository.Exists(id))
+            {
+                return NotFound();
+            }
+            else if (!await _housingRepository.IsOwnedByBrokerFirm(id, brokerFirmId))
+			{
+                return BadRequest();
+            }            
+
             await _housingRepository.DeleteAsync(id);
             return Ok();
         }
 
-        /// <summary>
-        /// An API endpoint for searching housing objects. 
-        /// </summary>
-        /// <param name="brokerId">The ID of the broker associated with the housing.</param>
-        /// <param name="municipalityId">An optional municipality filter.</param>
-        /// <returns>An embedded collection of <see cref="HousingDto"/>.</returns>
-        /// <!-- Author: Jimmie -->
-        /// <!-- Co Authors: -->
-        [HttpGet]
+		/// <summary>
+		/// An API endpoint for searching housing objects. 
+		/// </summary>
+		/// <param name="brokerFirmId">The ID of the broker firm associated with the housing.</param>
+		/// <param name="brokerId">The ID of the broker associated with the housing.</param>
+		/// <param name="municipalityId">An optional municipality filter.</param>
+		/// <returns>An embedded collection of <see cref="HousingDto"/>.</returns>
+		/// <!-- Author: Jimmie -->
+		/// <!-- Co Authors: -->
+		[HttpGet]
         [ProducesResponseType<HousingDto>(StatusCodes.Status200OK)]
-        public async Task<ActionResult<IEnumerable<HousingDto>>> Get([Required] int brokerId, int? municipalityId = null)
+        public async Task<ActionResult<IEnumerable<HousingDto>>> Get([Required] int brokerFirmId, int? brokerId, int? municipalityId = null)
         {
-            var housings = (await _housingRepository.GetAllHousingAsync(municipalityId, brokerId))
+            var housings = (await _housingRepository.GetAllHousingAsync(municipalityId, brokerId, brokerFirmId))
                 .Select(x => _mapper.Map<HousingDto>(x))
                 .ToList();
 
@@ -95,19 +108,23 @@ namespace FribergFastigheter.Server.Controllers.BrokerFirmApi
         /// An API endpoint for fetching a housing object. 
         /// </summary>
         /// <param name="id">The ID of the housing to fetch.</param>
-        /// <param name="brokerId">The ID of the broker associated with the housing.</param>
+        /// <param name="brokerFirmId">The ID of the broker firm associated with the housing.</param>
         /// <returns>An embedded collection of <see cref="HousingDto"/>.</returns>
         /// <!-- Author: Jimmie -->
         /// <!-- Co Authors: -->
         [HttpGet("{id:int}")]
         [ProducesResponseType<HousingDto>(StatusCodes.Status200OK)]
-        public async Task<ActionResult<IEnumerable<HousingDto>>> GetById(int id, [Required] int brokerId)
+        public async Task<ActionResult<IEnumerable<HousingDto>>> GetById(int id, [Required] int brokerFirmId)
         {
-            var housing = await _housingRepository.GetHousingByIdAsync(id, brokerId);
+            var housing = await _housingRepository.GetHousingByIdAsync(id);
 
             if (housing == null)
             {
                 return NotFound();
+            }
+            else if (housing.BrokerFirm.BrokerFirmId != brokerFirmId)
+            {
+                return BadRequest();
             }
 
             var result = _mapper.Map<HousingDto>(housing);
@@ -116,41 +133,43 @@ namespace FribergFastigheter.Server.Controllers.BrokerFirmApi
             return Ok(result);
         }
 
-        /// <summary>
-        /// An API endpoint for creating housing objects. 
-        /// </summary>
-        /// <param name="brokerId">The ID of the broker associated with the new housing.</param>
-        /// <param name="housingDto">The serialized DTO object.</param>
-        [HttpPost]
-        public async Task<ActionResult> Post([FromBody] CreateHousingDto housingDto)
+		/// <summary>
+		/// An API endpoint for creating housing objects. 
+		/// </summary>
+		/// param name="brokerFirmId">The ID of the broker firm associated with the housing.</param>
+		/// <param name="housingDto">The serialized DTO object.</param>
+		[HttpPost]
+        public async Task<ActionResult> Post([Required] int brokerFirmId, [FromBody] CreateHousingDto housingDto)
         {
-            var newHousing = _mapper.Map<Housing>(housingDto);
+			if (brokerFirmId != housingDto.BrokerFirmId)
+			{
+				return BadRequest();
+			}
 
-            return Ok();
-
+			var newHousing = _mapper.Map<Housing>(housingDto);
             await _housingRepository.AddAsync(newHousing);
             return Ok();
         }
 
-        /// <summary>
-        /// An API endpoint for updating housing objects. 
-        /// </summary>
-        /// <param name="id">The ID of the housing object to update.</param>
-        /// <param name="brokerId">The ID of the broker associated with the new housing.</param>
-        /// <param name="housingDto">The serialized DTO object.</param>
-        [HttpPut("{id:int}")]
-        public async Task<ActionResult> Put(int id, [FromBody] UpdateHousingDto housingDto)
+		/// <summary>
+		/// An API endpoint for updating housing objects. 
+		/// </summary>
+		/// <param name="id">The ID of the housing object to update.</param>
+		/// <param name="brokerFirmId">The ID of the broker firm associated with the housing.</param>
+		/// <param name="housingDto">The serialized DTO object.</param>
+		[HttpPut("{id:int}")]
+        public async Task<ActionResult> Put(int id, [Required] int brokerFirmId, [FromBody] UpdateHousingDto housingDto)
         {
             if (id != housingDto.HousingId)
             {
                 return BadRequest();
             }
+            else if (!(await _housingRepository.IsOwnedByBrokerFirm(id, brokerFirmId)))
+            {
+                return BadRequest();
+            }
 
-            var housing = _mapper.Map<Housing>(housingDto);
-
-            return Ok();
-
-            await _housingRepository.UpdateAsync(housing);
+            await _housingRepository.UpdateAsync(_mapper.Map<Housing>(housingDto));
             return Ok();
         }
